增大连接agent时间，修改表字段类型

+++ src/db/task_queue_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Enum, BigInteger
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Enum, BigInteger, Float
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 from src.utils.base import Base
@@ -11,7 +11,7 @@ class TaskQueueModel(Base):
     patient_id = Column(Integer, ForeignKey('patients.id'), nullable=False, comment='病人ID')
     conversation_id = Column(BigInteger, ForeignKey('conversations.id'), nullable=False, comment='会话ID')
     status = Column(Enum('queued', 'running', 'completed', 'failed'), nullable=False, default='queued', comment='任务状态')
-    estimated_time = Column(Integer, nullable=False, default=10, comment='任务单独执行所需时间（秒）')
+    estimated_time = Column(Float, nullable=False, default=10.0, comment='任务单独执行所需时间（秒）')
     created_at = Column(DateTime, default=lambda: datetime.now(), comment='创建时间')
     updated_at = Column(DateTime, default=lambda: datetime.now(), onupdate=lambda: datetime.now(), comment='更新时间')
     started_at = Column(DateTime, nullable=True, comment='开始执行时间')
